+++ src/test/static/polymer2/test-element-12.js
@@ -18,7 +18,7 @@ class TestElement extends Polymer.Element {
     return [
       'let let let parseError',
       'foo',
-      'foo(bar)',
+      'foo(bar)'
     ];
   }
 }